pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "ebtools"
dynamic = ["version"]
description = "EarthBound/Mother 2 ROM asset extractor"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0", "cyclopts>=3.6", "Pillow>=9.0", "pydantic>=2.0"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "ebtools/_version.py"

[project.scripts]
ebtools = "ebtools.cli:app"

[tool.ruff]
target-version = "py310"

exclude = [
    "migrations",
    "__pycache__",
    "manage.py",
    "settings.py",
    "env",
    ".env",
    "venv",
    ".venv",
    "ebtools/_version.py",
]

line-length = 120

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "D",    # pydocstyle
    "E",    # Error
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "PGH",  # pygrep-hooks
    "PTH",  # flake8-use-pathlib
    "Q",    # flake8-quotes
    "S",    # bandit
    "SIM",  # flake8-simplify
    "TRY",  # tryceratops
    "UP",   # pyupgrade
    "W",    # Warning
    "YTT",  # flake8-2020
]

ignore = [
    "B905",    # zip strict=True; remove once python <3.10 support is dropped.
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D401",
    "E402",
    "E501",
    "F401",
    "N815",    # mixedCase class variables; dataclass fields match YAML keys
    "S101",    # use of assert; used for internal validation
    "TRY002",  # bare Exception; acceptable for this codebase
    "TRY003",  # Avoid specifying messages outside exception class; overly strict
]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "chr",
    "typer.Argument",
    "typer.Option",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "S101",    # use of "assert"
    "S102",    # use of "exec"
    "S106",    # possible hardcoded password.
    "PGH001",  # use of "eval"
]

[tool.ruff.lint.pep8-naming]
staticmethod-decorators = [
    "pydantic.validator",
    "pydantic.root_validator",
]
